test(metrics): run service-module metric tests in unit lane
The integration lane runs tests/service inside the agent-svc container,
which lacks curl_cffi and playwright. Move the eight stage-metric tests
that import scraper.app (curl_cffi) and browser_svc.app (playwright) into
tests/unit so they only run in Fast Tests, where those deps are available.
Test bodies are unchanged.
